+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 
  - Added scope-aware completion.
+ - Added scope lookup by path method in the indexer.
  - Added `--version` management for sv-indexer (same version as LSP)
  
 ## Changed
@@ -14,6 +15,7 @@
 ## Fixed
 
  - Avoid spurious errors in log due to unknown command `textDocument/didClose`
+ - Fixed `diplomat-server.list-symbols` to handle design path as an input (#17)
 
 # 0.2.1